+++ argilla-server/src/argilla_server/api/handlers/v1/datasets/records.py
@@ -34,6 +34,7 @@ from argilla_server.api.schemas.v1.records import (
     SearchRecordsQuery,
     SearchRecordsResult,
     TermsFilter,
+    SEARCH_MAX_SIMILARITY_SEARCH_RESULT,
 )
 from argilla_server.api.schemas.v1.records import Record as RecordSchema
 from argilla_server.api.schemas.v1.responses import ResponseFilterScope
@@ -44,9 +45,9 @@ from argilla_server.api.schemas.v1.suggestions import (
     SuggestionFilterScope,
 )
 from argilla_server.api.handlers.v1.workspaces import list_workspace_users
-from argilla_server.contexts import datasets, search
+from argilla_server.contexts import datasets, search, records
 from argilla_server.database import get_async_db
-from argilla_server.enums import RecordSortField, ResponseStatusFilter
+from argilla_server.enums import RecordSortField
 from argilla_server.errors.future import MissingVectorError, NotFoundError, UnprocessableEntityError
 from argilla_server.errors.future.base_errors import MISSING_VECTOR_ERROR_CODE
 from argilla_server.models import Dataset, Field, Record, User, VectorSettings, Response, Question, Suggestion
@@ -54,7 +55,6 @@ from argilla_server.search_engine import (
     AndFilter,
     SearchEngine,
     SearchResponses,
-    UserResponseStatusFilter,
     get_search_engine,
 )
 from argilla_server.security import auth
@@ -71,7 +71,6 @@ parse_record_include_param = parse_query_param(
 )
 
 router = APIRouter()
-
 
 async def _filter_records_using_search_engine(
     db: "AsyncSession",
@@ -102,14 +101,13 @@ async def _filter_records_using_search_engine(
         search_responses.total,
     )
 
-
 def _to_search_engine_filter_scope(scope: FilterScope, user: Optional[User]) -> search_engine.FilterScope:
     if isinstance(scope, RecordFilterScope):
         return search_engine.RecordFilterScope(property=scope.property)
     elif isinstance(scope, MetadataFilterScope):
         return search_engine.MetadataFilterScope(metadata_property=scope.metadata_property)
     elif isinstance(scope, SuggestionFilterScope):
-        return search_engine.SuggestionFilterScope(question=scope.question, property=scope.property)
+        return search_engine.SuggestionFilterScope(question=scope.question, property=str(scope.property))
     elif isinstance(scope, ResponseFilterScope):
         return search_engine.ResponseFilterScope(question=scope.question, property=scope.property, user=user)
     else:
@@ -199,13 +197,20 @@ async def _get_search_responses(
             "record": record,
             "query": text_query,
             "order": vector_query.order,
-            "max_results": limit,
+            "max_results": min(limit + offset, SEARCH_MAX_SIMILARITY_SEARCH_RESULT),
         }
 
         if filters:
             similarity_search_params["filter"] = _to_search_engine_filter(filters, user=user)
 
-        return await search_engine.similarity_search(**similarity_search_params)
+        if offset >= similarity_search_params["max_results"]:
+            return SearchResponses(items=[], total=0)
+
+        responses = await search_engine.similarity_search(**similarity_search_params)
+        responses.items = responses.items[offset:]
+
+        return responses
+
     else:
         search_params = {
             "dataset": dataset,
@@ -223,18 +228,6 @@ async def _get_search_responses(
             search_params["sort"] = _to_search_engine_sort(sort, user=user)
 
         return await search_engine.search(**search_params)
-
-
-async def _build_response_status_filter_for_search(
-    response_statuses: Optional[List[ResponseStatusFilter]] = None, user: Optional[User] = None
-) -> Optional[UserResponseStatusFilter]:
-    user_response_status_filter = None
-
-    if response_statuses:
-        # TODO(@frascuchon): user response and status responses should be split into different filter types
-        user_response_status_filter = UserResponseStatusFilter(user=user, statuses=response_statuses)
-
-    return user_response_status_filter
 
 
 async def _validate_search_records_query(db: "AsyncSession", query: SearchRecordsQuery, dataset: Dataset):
@@ -302,7 +295,6 @@ def generate_suggestions_from_response(
 async def list_dataset_records(
     *,
     db: AsyncSession = Depends(get_async_db),
-    search_engine: SearchEngine = Depends(get_search_engine),
     dataset_id: UUID,
     include: Optional[RecordIncludeParam] = Depends(parse_record_include_param),
     offset: int = 0,
@@ -310,19 +302,27 @@ async def list_dataset_records(
     current_user: User = Security(auth.get_current_user),
 ):
     dataset = await Dataset.get_or_raise(db, dataset_id)
-
     await authorize(current_user, DatasetPolicy.list_records_with_all_responses(dataset))
 
-    records, total = await _filter_records_using_search_engine(
-        db,
-        search_engine,
-        dataset=dataset,
-        limit=limit,
-        offset=offset,
-        include=include,
+    include_args = (
+        dict(
+            with_responses=include.with_responses,
+            with_suggestions=include.with_suggestions,
+            with_vectors=include.with_all_vectors or include.vectors,
+        )
+        if include
+        else {}
     )
 
-    return Records(items=records, total=total)
+    dataset_records, total = await records.list_dataset_records(
+        db=db,
+        dataset_id=dataset.id,
+        offset=offset,
+        limit=limit,
+        **include_args,
+    )
+
+    return Records(items=dataset_records, total=total)
 
 
 @router.delete("/datasets/{dataset_id}/records", status_code=status.HTTP_204_NO_CONTENT)